+++ Code/src/WeatherStationProject.Dashboard.RainfallService/Services/IRainfallService.cs
@@ -5,6 +5,6 @@ namespace WeatherStationProject.Dashboard.RainfallService.Services
 {
     public interface IRainfallService
     {
-        Task<int> GetRainfallDuringTime(DateTime since, DateTime until);
+        Task<decimal> GetRainfallDuringTime(DateTime since, DateTime until);
     }
 }
